fixed intptr_t to int, thanks to greg
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@982 a1c6a512-1295-4272-9138-f99709370657

+++ firmware/common/dir.h
@@ -57,7 +57,7 @@ typedef struct {
 typedef struct DIRtag
 {
     struct dirent   fd;
-    intptr_t        handle;
+    int             handle;
 } DIR;
 #endif //   WIN32
 #endif // SIMULATOR
